[plugin] split arg parsing and logic.
